requirements.txt
# Face Recognition System - Streamlit Requirements

# Core deep learning framework
tensorflow>=2.10.0
keras>=2.10.0

# Image processing
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.6.0

# Streamlit and UI components
streamlit>=1.28.0
streamlit-webrtc>=0.45.0

# Additional utilities
pandas>=1.5.0
av>=10.0.0  # For video processing
aiortc>=1.3.0  # For WebRTC support

# Optional: For better performance
# streamlit-option-menu>=0.3.0
# plotly>=5.0.0  # For advanced visualizations